+++ app/lib/src/pages/profile_page/profile_page.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'package:talk_pilot/src/models/user_model.dart';
+import 'package:talk_pilot/src/pages/profile_page/widgets/stats_summary_card.dart';
 import 'package:talk_pilot/src/provider/user_provider.dart';
 import 'package:talk_pilot/src/components/toast_message.dart';
 
@@ -159,151 +160,6 @@ class _ProfilePageState extends State<ProfilePage> {
           ],
         ),
       ),
-    );
-  }
-}
-
-class StatsSummaryCard extends StatelessWidget {
-  final int completedCount;
-  final double averageScore;
-  final int averageCPM;
-  final double targetScore;
-
-  const StatsSummaryCard({
-    super.key,
-    required this.completedCount,
-    required this.averageScore,
-    required this.averageCPM,
-    required this.targetScore,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      color: Colors.white,
-      elevation: 3,
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-      child: Padding(
-        padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 8),
-        child: Column(
-          children: [
-            SizedBox(
-              height: 104,
-              child: Row(
-                children: [
-                  Expanded(
-                    child: _StatItem(
-                      icon: Icons.check_circle,
-                      value: '$completedCount',
-                      label: '완료한 발표',
-                    ),
-                  ),
-                  Container(
-                    width: 1,
-                    height: double.infinity,
-                    color: Colors.grey.shade300,
-                  ),
-                  Expanded(
-                    child: _StatItem(
-                      icon: Icons.star,
-                      value: averageScore.toStringAsFixed(1),
-                      label: '평균 점수',
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            const SizedBox(height: 8),
-            Row(
-              children: [
-                Expanded(
-                  child: Container(
-                    height: 1,
-                    color: Colors.grey.shade300,
-                    margin: const EdgeInsets.only(right: 4),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                Expanded(
-                  child: Container(
-                    height: 1,
-                    color: Colors.grey.shade300,
-                    margin: const EdgeInsets.only(left: 4),
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 8),
-            SizedBox(
-              height: 104,
-              child: Row(
-                children: [
-                  Expanded(
-                    child: _StatItem(
-                      icon: Icons.speed,
-                      value: '$averageCPM',
-                      label: '평균 CPM',
-                    ),
-                  ),
-                  Container(
-                    width: 1,
-                    height: double.infinity,
-                    color: Colors.grey.shade300,
-                  ),
-                  Expanded(
-                    child: _StatItem(
-                      icon: Icons.flag,
-                      value: targetScore.toStringAsFixed(1),
-                      label: '목표 점수',
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-
-class _StatItem extends StatelessWidget {
-  final IconData icon;
-  final String value;
-  final String label;
-
-  const _StatItem({
-    required this.icon,
-    required this.value,
-    required this.label,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Icon(icon, size: 26, color: Colors.deepPurple), // 아이콘 키움
-        const SizedBox(height: 6),
-        Text(
-          value,
-          style: const TextStyle(
-            fontSize: 18,
-            fontWeight: FontWeight.bold,
-            color: Colors.black87,
-          ),
-        ),
-        const SizedBox(height: 4),
-        Text(
-          label,
-          style: const TextStyle(
-            fontSize: 14,
-            color: Colors.grey,
-          ),
-          textAlign: TextAlign.center,
-        ),
-      ],
     );
   }
 }
